fixed some bugs

+++ db_api/app/User.py
@@ -1,4 +1,5 @@
-from app import app, cursor
+# -*- coding: utf-8 -*-
+from app import app, cursor, conn
 from flask import request, jsonify
 from functions import *
 import json
@@ -6,7 +7,7 @@ import Forum
 
 @app.route("/db/api/user/create/", methods = ['POST'])
 def createUser():
-    logging.info("================USER CREATION")
+
     try:
         email    = request.json["email"]
     except:
@@ -16,29 +17,21 @@ def createUser():
         name     = request.json["name"]
     except:
         name     = None
-        logging.info("NAME IS None")
-        logging.info("_____________________")
     try:
         username = request.json["username"]
     except:
         username = None
-        logging.info("USERNAME IS None")
-        logging.info("_____________________")
     try:
         about    = request.json["about"]
     except:
         about    = None
-        logging.info("ABOUT IS None")
-        logging.info("_____________________")
 
     if (not isString([name, email, username, about])):
-        logging.info("one or more params not string")
         return json.dumps({"code": 2, "response": error_messages[2]})
 
     sql = "SELECT idUser FROM User WHERE email = %s"
     cursor.execute(sql, [email])
-    if (cursor.fetchone() != None):
-        logging.info(email + " is already exists")
+    if cursor.fetchone() != None:
         return json.dumps({"code": 5, "response": error_messages[5]})
 
     isAnonymous = False
@@ -47,6 +40,8 @@ def createUser():
 
     sql = "INSERT INTO User(username, email, name, about, isAnonymous) VALUES (%s, %s, %s, %s, %s)"
     cursor.execute(sql, [username, email, name, about, isAnonymous])
+
+    conn.commit()
 
     sql = "SELECT MAX(idUser) FROM User"
     cursor.execute(sql)
@@ -61,7 +56,6 @@ def createUser():
     data['username'] = username
     answer = {"code": 0, "response": data}
     response = json.dumps(answer)
-    logging.info("================SUCCESSFUL USER CREATION\n")
     return response
 
 @app.route("/db/api/user/details/", methods = ['GET'])
@@ -70,7 +64,6 @@ def userDetails():
         email = request.args.get("user")
     except:
         return json.dumps({"code": 2, "response": error_messages[2]})
-    logging.info(email)
     if (not isString([email])):
         return json.dumps({"code": 2, "response": error_messages[2]})
 
@@ -83,28 +76,27 @@ def userDetails():
 
 @app.route("/db/api/user/follow/", methods = ['POST'])
 def follow():
-    logging.info("================USER FOLLOW========================")
     try:
         follower = request.json["follower"]
         followee = request.json["followee"]
     except:
         return json.dumps({"code": 2, "response": error_messages[2]})
-    try:
-        idFollower = getUserIdByEmail(follower)
-        idFollowee = getUserIdByEmail(followee)
-    except:
+
+    idFollower = getUserIdByEmail(follower)
+    idFollowee = getUserIdByEmail(followee)
+
+    if (idFollower is None or idFollowee is None):
         return json.dumps({"code": 1, "response": error_messages[1]})
 
-    logging.info("  follower : " + str(idFollower))
-    logging.info("  followee : " + str(idFollowee))
-
     sql = "INSERT INTO Follower(idFollower, idFollowing) VALUES(%s, %s)"
+
 
     try:
         cursor.execute(sql, [idFollower, idFollowee])
     except:
         json.dumps({"code": 5, "response": error_messages[5]})
 
+    conn.commit()
 
     emailsFollowers = getFollowerEmails(idFollower, None, None, None)
     emailsFollowing = getFollowingEmails(idFollower, None, None, None)
@@ -113,15 +105,11 @@ def follow():
     userInfo["followers"] = emailsFollowers
     userInfo["following"] = emailsFollowing
     userInfo["subscriptions"] = []
-    logging.info("Response : ")
-    logging.info(userInfo)
     response = json.dumps({"code": 0, "response": userInfo})
-    logging.info("================USER FOLLOW SUCCESS================\n")
     return response
 
 @app.route("/db/api/user/listFollowers/", methods = ['GET'])
 def listFollowers():
-    logging.info("================USER listFollowers========================")
     try:
         user = request.args.get("user")
     except:
@@ -131,7 +119,6 @@ def listFollowers():
     order = getOptionalGetParameterOrDefault(request.args, "order", "desc")
     since = getOptionalGetParameterOrDefault(request.args, "since_id", None)
 
-    logging.info("user : " + user)
 
     user_id = getUserIdByEmail(user)
     if not user_id:
@@ -144,15 +131,16 @@ def listFollowers():
         user_by_email = getUserInfoByEmail(email)
         answer.append((user_by_email))
 
+    # if not answer:
+    #     print("empty")
+    #     return json.dumps({"code": 1, "response": error_messages[2]})
+
+
     response = json.dumps({ "code": 0, "response": answer })
-    logging.info("Response : ")
-    logging.info(response)
-    logging.info("================USER listFollowers END====================\n")
     return response
 
 @app.route("/db/api/user/listFollowing/", methods = ['GET'])
 def listFollowing():
-    logging.info("================USER listFollowing========================")
     try:
         user = request.args.get("user")
     except:
@@ -162,7 +150,6 @@ def listFollowing():
     order = getOptionalGetParameterOrDefault(request.args, "order", "desc")
     since = getOptionalGetParameterOrDefault(request.args, "since_id", None)
 
-    logging.info("  User : " + user)
 
     user_id = getUserIdByEmail(user)
     if not user_id:
@@ -171,19 +158,21 @@ def listFollowing():
     emails = getFollowingEmails(user_id, since, order, limit)
 
     answer = []
+
+
     for email in emails:
         user_by_email = getUserInfoByEmail(email)
         answer.append((user_by_email))
 
+    # if not answer:
+    #     print("empty 1")
+    #     return json.dumps({"code": 1, "response": error_messages[2]})
+
     response = json.dumps({ "code": 0, "response": answer })
-    logging.info("Response : ")
-    logging.info(response)
-    logging.info("================USER listFollowing END====================\n")
     return response
 
 @app.route("/db/api/user/listPosts/", methods = ['GET'])
 def userListPosts():
-    logging.info("================USER LISTPOSTS================")
     try:
         user = request.args.get("user")
     except:
@@ -210,60 +199,52 @@ def userListPosts():
         params.append(int(limit))
         sql += " LIMIT %s"
 
-    logging.info("  Final SQL    listPosts : " + sql)
-    logging.info("  Final PARAMS listPosts : " + str(params))
-
     cursor.execute(sql, params)
     result = getArrayDictFromDoubleDictionary(cursor.fetchall())
-    logging.info("  Results : ")
+
+    # if not result:
+    #     print("empty 2")
+    #     return json.dumps({"code": 1, "response": error_messages[2]})
 
     response = json.dumps({ "code": 0, "response": result})
-    logging.info("  Response : " + response)
-    logging.info("================USER LISTPOSTS END============")
     return response
 
 @app.route("/db/api/user/unfollow/", methods = ['POST'])
 def unfollow():
-    logging.info("================USER UNFOLLOW========================")
     try:
         follower = request.json["follower"]
         followee = request.json["followee"]
     except:
         return json.dumps({"code": 2, "response": error_messages[2]})
-    try:
-        idFollower = getUserIdByEmail(follower)
-        idFollowee = getUserIdByEmail(followee)
-    except:
-        return json.dumps({"code": 1, "response": error_messages[1]})
 
-    logging.info("  follower : " + str(idFollower))
-    logging.info("  followee : " + str(idFollowee))
+    idFollower = getUserIdByEmail(follower)
+    idFollowee = getUserIdByEmail(followee)
+
+    if (idFollower is None or idFollowee is None):
+        return json.dumps({"code": 1, "response": error_messages[1]})
 
     sql = "DELETE FROM Follower WHERE idFollower = %s AND idFollowing = %s"
     cursor.execute(sql, [idFollower, idFollowee])
+    conn.commit()
     answer = getUserInfoByID(idFollower)
 
     response = json.dumps({"code": 0, "response": answer})
-    logging.info("================USER UNFOLLOW END====================")
     return response
 
 @app.route("/db/api/user/updateProfile/", methods = ['POST'])
 def updateProfile():
-    logging.info("=======================UPDATING PROFILE========================")
     try:
         about = request.json["about"]
         user = request.json["user"]
         name = request.json["name"]
     except:
-        logging.error("  no required fields")
         return json.dumps({"code": 2, "response": error_messages[2]})
     sql = "UPDATE User SET about = %s, name = %s WHERE email = %s"
     cursor.execute(sql, [about, name, user])
+    conn.commit()
     result = getUserInfoByEmail(user)
 
     response = json.dumps({"code": 0, "response": result})
-    logging.info("  Result : " + response)
-    logging.info("=======================UPDATING PROFILE SUCCESSFUL=============")
     return response
 
 def getUserInfoByEmail(email):
@@ -288,12 +269,13 @@ def getUserInfoByEmail(email):
         return None
 
 def getUserIdByEmail(email):
-    logging.info("      getting user by email : ")
     sql = "SELECT idUser FROM User WHERE email = %s"
     cursor.execute(sql, [email])
-    q_result = cursor.fetchone()[0]
-    logging.info("      getted user by email : " + str(q_result))
-    return q_result
+    q_result = cursor.fetchone()
+    if q_result != None:
+        return q_result[0]
+    else:
+        return None
 
 def getUserInfoByID(id):
     sql = "SELECT * FROM User WHERE idUser = %s"
@@ -328,7 +310,6 @@ def getUserEmailByID(id):
         return None
 
 def getFollowerEmails(idUser, since, order, limit):
-    logging.info("")
     sql = "SELECT U.email FROM Follower F INNER JOIN User U ON F.idFollower = U.idUser WHERE F.idFollowing = %s"
     params = [idUser]
     if since:
@@ -341,12 +322,8 @@ def getFollowerEmails(idUser, since, order, limit):
     if limit:
         sql += " LIMIT %s"
         params.append(int(limit))
-    logging.info("      Final SQL    followerEmails : " + sql)
-    logging.info("      Final PARAMS followerEmails : " + str(params))
     cursor.execute(sql, params)
     emails = getArrayEmailsFromDoubleDictionary(cursor.fetchall())
-    logging.info("      EMAILS FOLLOWERS OF USER (" + str(idUser) + ") : " + str(emails))
-    logging.info("")
     return emails
 
 def getFollowingEmails(idUser, since, order, limit):
@@ -362,8 +339,6 @@ def getFollowingEmails(idUser, since, order, limit):
     if limit:
         sql += " LIMIT %s"
         params.append(int(limit))
-    logging.info("      Final SQL    followingEmails : " + sql)
-    logging.info("      Final PARAMS followingEmails : " + str(params))
     cursor.execute(sql, params)
     emails = getArrayEmailsFromDoubleDictionary(cursor.fetchall())
     logging.info("      EMAILS FOLLOWING OF USER (" + str(idUser) + ") : " + str(emails))
